+++ CHANGELOG.md
@@ -7,6 +7,7 @@
 - POTENTIALLY BREAKING: Fixed a typing bug where the `totalSupply` field in an `NftContract` should have type `string` instead of `number`.
 - Updated the `Nft` class to include the contract metadata in the `Nft.contract` field.
 - Added commonly used utility methods from ethers.js into a top-level `Utils` export.
+- Added the `NftNamespace.verifyNftOwnership()` method to replace the deprecated `checkNftOwnership()` method. 
 - Added `Contract` and `ContractFactory` exports from ethers into the top-level exports.  
 
 ### Minor Changes